#70 照片上传下载调通

+++ seen-common/seen-core/src/main/java/io/github/seenings/common/model/UserIdAndToken.java
@@ -6,6 +6,12 @@ import lombok.Data;
 @Builder
 @Data
 public class UserIdAndToken {
-    private int userId;
+    /**
+     * 用户ID
+     */
+    private long userId;
+    /**
+     * 令牌
+     */
     private String token;
 }
